requirements.txt
# Core
streamlit
python-dotenv
PyPDF2

# OpenAI y LangChain
openai>=1.0.0
langchain
langchain-openai
langchain-community

# Vector store (FAISS)
faiss-cpu
